% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_table}
\alias{fetch_query}
\title{Query the AAEDB}
\usage{
fetch_table(x, schema = "aquatic_data")

fetch_query(query)
}
\arguments{
\item{x}{a character specifying an individual table in the AAEDB}

\item{schema}{schema in which \code{x} is found. Defaults to
\code{"aquatic_data"}}
}
\description{
\code{fetch_table} and \code{fetch_query} represent two
  ways to interact with the AAEDB. \code{fetch_table} provides access
  to existing flat tables in the database, whereas \code{fetch_query}
  allows users to compute custom queries.
  
  Both functions require credentials to access the AAEDB, plus appropriate
  access and credentials to the relevant VPN.
  
  \code{fetch_query} is NOT CURRENTLY IMPLEMENTED.
}
\examples{
# to add


}
