% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a QA/QC report on a data set for the AAEDB}
\usage{
generate_report(data, filename = NULL, output = "template.html")
}
\arguments{
\item{data}{a data set to be imported to the AAEDB, provided
as a \code{matrix}, \code{data.frame}, or \code{tibble}}

\item{output}{name of output file, including path. By default,
will be saved as `template.html` in the top level of the
current working directory}
}
\description{
Generate a QA/QC report on a data set for the AAEDB
}
\examples{
# to add
}
